+++ database/migrations/2025_03_05_170937_create_evrak_antrepo_giris_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('vekaletFirmaKisiAdi');  // firma tablosundan
             $table->string('urunAdi');
             $table->string('gtipNo');
-            $table->integer('urunKG');
+            $table->decimal('urunKG',10,3);
             $table->string('sevkUlke');
             $table->string('orjinUlke');
             $table->string('aracPlaka');
